회원탈퇴 확인 팝업 추가
이메일 중복확인 팝업 추가

+++ app/structure/lib/viewModel/my_page/delete_user_view_model.dart
@@ -1,6 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
+import 'package:structure/components/custom_dialog.dart';
 import 'package:structure/config/labels.dart';
 import 'package:structure/config/pallete.dart';
 import 'package:structure/dataSource/local_data_source.dart';
@@ -17,6 +18,7 @@ class DeleteUserViewModel with ChangeNotifier {
   TextEditingController password = TextEditingController();
 
   bool _isValidPw = false;
+  bool _confirmDelete = false;
 
   /// 기존 비밀번호 유효성 검사
   String? pwValidate(String? value) {
@@ -40,46 +42,64 @@ class DeleteUserViewModel with ChangeNotifier {
 
   late BuildContext _context;
 
+  ///회원탈퇴 확인 dialog
+  Future<void> _showDeleteIdDialog(dynamic response) async {
+    showDeleteIdDialog(_context, popDialog, popDialogConfirm);
+  }
+
+  void popDialog() {
+    _context.pop();
+    _context.pop();
+  }
+
+  void popDialogConfirm() {
+    _confirmDelete = true;
+    _context.pop();
+  }
+
   /// 회원 탈퇴 함수
   Future<void> deleteUser(BuildContext context) async {
     _context = context;
-    isLoading = true;
-    notifyListeners();
-    try {
-      // 기존 firebase user 정보 불러오기 (현재 로그인된 유저)
-      User? user = FirebaseAuth.instance.currentUser;
-      if (user != null) {
-        // 비밀번호 변경 전 firebase에 재인증 필요
-        await FirebaseAuth.instance.currentUser?.reauthenticateWithCredential(
-          EmailAuthProvider.credential(
-            email: userModel.userId!,
-            password: password.text,
-          ),
-        );
+    _showDeleteIdDialog(context);
+    if (_confirmDelete) {
+      isLoading = true;
+      notifyListeners();
+      try {
+        // 기존 firebase user 정보 불러오기 (현재 로그인된 유저)
+        User? user = FirebaseAuth.instance.currentUser;
+        if (user != null) {
+          // 비밀번호 변경 전 firebase에 재인증 필요
+          await FirebaseAuth.instance.currentUser?.reauthenticateWithCredential(
+            EmailAuthProvider.credential(
+              email: userModel.userId!,
+              password: password.text,
+            ),
+          );
 
-        // Firebase 유저 삭제
-        // await user.delete();
+          // Firebase 유저 삭제
+          // await user.delete();
 
-        // DB에서 유저 삭제 API 호출
-        final response = await RemoteDataSource.deleteUser(userModel.userId!);
-        if (response == null) {
-          throw Error();
+          // DB에서 유저 삭제 API 호출
+          final response = await RemoteDataSource.deleteUser(userModel.userId!);
+          if (response == null) {
+            throw Error();
+          }
+
+          _success();
+        } else {
+          print('User does not exist.');
         }
-
-        _success();
-      } else {
-        print('User does not exist.');
+      } on FirebaseException catch (e) {
+        print('error: ${e.code}');
+        if (e.code == 'wrong-password') {
+          _showAlert(Labels.pwdNotSame); // 기존 비밀번호가 틀리면 alert 생성
+        } else {
+          _showAlert('오류가 발생했습니다.');
+        }
       }
-    } on FirebaseException catch (e) {
-      print('error: ${e.code}');
-      if (e.code == 'wrong-password') {
-        _showAlert(Labels.pwdNotSame); // 기존 비밀번호가 틀리면 alert 생성
-      } else {
-        _showAlert('오류가 발생했습니다.');
-      }
+      isLoading = false;
+      notifyListeners();
     }
-    isLoading = false;
-    notifyListeners();
   }
 
   /// 오류 snackbar
